backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0

